Misc tweaks and code cleanup
* Imagine: Simplify Viewport class and move it out of Gfx namespace
* Imagine: Store a default viewport rectangle in Renderer's per-window data so it doesn't need to be set every frame when using the same Drawable
* Imagine: Add noreturn attribute to bug_doExit() and remove various return statements in unreachable code
* Imagine: Add util/enum.hh and incorporate wise_enum for additional functionality
* Imagine: Refactor Pixmap class to PixmapView and MutablePixmapView to allow for const image data
* EmuFramework: Use std::span for defaultConfigsForInputMap() to avoid need for out parameter
* EmuFramework: Fix color format of VideoImageOverlay's CRT RGB texture
* EmuFramework: Move EmuViewController into main window's app data

+++ EmuFramework/src/vcontrols/VControllerButton.cc
@@ -190,7 +190,7 @@ std::array<int, 2> VControllerButtonGroup::findButtonIndices(IG::WP windowPos) c
 	std::array<int, 2> btnOut{-1, -1};
 	if(state() == VControllerState::OFF)
 		return btnOut;
-	for(unsigned count = 0; auto &b : buttons())
+	for(size_t count = 0; auto &b : buttons())
 	{
 		if(b.isEnabled() && b.realBounds().overlaps(windowPos))
 		{
